jiahn Entity작업 - portfolio, price, order

+++ src/main/java/hot/member/domain/Price.java
@@ -5,6 +5,7 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import javax.persistence.OneToMany;
 import javax.persistence.Table;
 
 import lombok.AllArgsConstructor;
@@ -23,7 +24,7 @@ public class Price {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "price_no")
-	private Long priceNo;
+	private int priceNo;
 	
 	@Column(name = "price_amount")
 	private int priceAmount;
